pyproject.toml
[tool.poetry]
name = "marvel-user-api"
version = "0.1.0"
description = ""
authors = ["jemmanespavi"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
motor = "^2.5.1"
bcrypt = "^3.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
